Fixes CRConnect #414 and #415
makes script editor wider; makes font/style more legible; makes dropdown script list opaque. noted limitations: (1) it is a PITA to get script editor to resize out to the left, so i just made it a big bigger overall; (2) the script list is stupid long now but it is a PITA to put a scrollbar on the menu and preserve the popout elements. I included a stackoverflow link in there in case its something to add in the future though